Created new signup related HTML and CSS, as well as login related HTML and CSS, and forget_password related HTML and CSS pages
Created signup and login pages associated with the initial signup_ond-login page, maintaining the same style as before, and also created a forgetpassword page based on the login page that allows users to retrieve their passwords. The links associated with it can also lead to each corresponding page. At the same time, a JS file with hidden password function was also created.